libft: Changed ft_count_digits to support numbers up to long long.

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: saaltone <saaltone@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/29 10:42:36 by saaltone          #+#    #+#             */
-/*   Updated: 2022/02/14 15:06:22 by saaltone         ###   ########.fr       */
+/*   Updated: 2022/02/15 13:47:06 by saaltone         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -104,7 +104,7 @@ void	ft_lstput(t_list **alst, t_list *new_el);
 void	ft_lstiter(t_list *lst, void (*f)(t_list *elem));
 t_list	*ft_lstmap(t_list *lst, t_list *(*f)(t_list *elem));
 int		ft_split_count(char *str, char c);
-int		ft_count_digits(int number);
+int		ft_count_digits(long long number);
 void	ft_putstr_color(char *str, char *color);
 double	ft_pow(double number, double power);
 int		ft_abs(int number);
